+++ web/app/index-external.ts
@@ -1,0 +1,4 @@
+/* This is the entry point file for the external modals */
+
+import "./ui/context-menu";
+import "tc-shared/ui/react-elements/external-modal/PopoutEntrypoint"