+++ ChangeLog.md
@@ -6,6 +6,7 @@
     - Added support for HTML encoded links (Example would be when copying from Edge the URL)
     - Enabled context menus for all clickable client tags
     - Allowing to drag client tags
+    - Fixed the context menu within popout windows for the web client
     
 * **08.12.20**
     - Fixed the permission editor not resolving unique ids